renamed some files and fixed typos
