+++ deep_learning4e.py
