+++ tests/test_deep_learning4e.py
@@ -1,6 +1,6 @@
 import pytest
 
-from DeepNeuralNet4e import *
+from deep_learning4e import *
 from learning4e import DataSet, grade_learner, err_ratio
 from keras.datasets import imdb
 import numpy as np
